Updates/Fixes to code
I added lines of code (13, 84, 85, 86, 87, 88) to correctly position the footer to extend to the left and right edges of the page as well as stay at the bottom of the page.

I also added a missing alt tag to the image for the formal blazer. This is done to provide an alternate text that is displayed if the image is unable to be shown.